aos-backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
pydantic==2.11.1
pydantic-settings==2.8.1

# Database
sqlalchemy==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pgvector==0.3.6

# Redis
redis==5.3.0

# Auth
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# AI / LLM
anthropic==0.52.0
openai==1.82.0
langchain==0.3.25
langchain-anthropic==0.3.18
langchain-openai==0.3.18
langgraph==0.4.1
tiktoken==0.9.0

# Document Processing
python-docx==1.1.2
openpyxl==3.1.5
PyPDF2==3.0.1

# HTTP
httpx==0.28.1
aiohttp==3.12.3

# Utilities
orjson==3.10.16
python-dateutil==2.9.0
phonenumbers==9.0.2
email-validator==2.2.0

# Observability
structlog==25.4.0
prometheus-client==0.22.1

# Testing
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.2.1
httpx==0.28.1

# File Storage
boto3==1.38.14
